# Make profiler work