<!--  Thanks for sending a pull request!  Here are some tips for you:

1. If this is your first time, please read our contributor guidelines: https://git.k8s.io/community/contributors/guide/first-contribution.md#your-first-contribution and developer guide https://git.k8s.io/community/contributors/devel/development.md#development-guide
2. Please label this pull request according to what type of issue you are addressing, especially if this is a release targeted pull request. For reference on required PR/issue labels, read here:
https://git.k8s.io/community/contributors/devel/sig-release/release.md#issuepr-kind-label
3. Ensure you have added or ran the appropriate tests for your PR: https://git.k8s.io/community/contributors/devel/sig-testing/testing.md
4. If you want *faster* PR reviews, read how: https://git.k8s.io/community/contributors/guide/pull-requests.md#best-practices-for-faster-reviews
5. If the PR is unfinished, see how to mark it: https://git.k8s.io/community/contributors/guide/pull-requests.md#marking-unfinished-pull-requests
-->

#### What type of PR is this?

/kind bug

#### What this PR does / why we need it:

Fix wzprof usage to enable profiling.

#### Which issue(s) this PR fixes:

none


#### Special notes for your reviewer:

A couple things were missing:

* The method wzprof.Profiling.Prepare needs to be called on the compiled module before instantiation.
* The wazero runtime needs to have debug info and custom sections enabled.

As you can see the method to get access to the CompiledModule from the profiler main function is a bit hacky. Feel free to edit.

#### Does this PR introduce a user-facing change?

```release-note
NONE
```

#### What are the benchmark results of this change?

N/A