@pelletier 2023-07-05
@codefromthecrypt here's the benchstat

```
goos: linux
goarch: amd64
pkg: sigs.k8s.io/kube-scheduler-wasm-extension/internal/e2e
cpu: AMD Ryzen 9 5950X 16-Core Processor
                                                        │ before.txt  │             after.txt             │
                                                        │   sec/op    │   sec/op     vs base              │
PluginFilter/noop-wat/params:_small-32                    325.1n ± 1%   322.6n ± 1%  -0.77% (p=0.002 n=6)
PluginFilter/noop-wat/params:_real-32                     326.3n ± 2%   322.1n ± 1%  -1.30% (p=0.026 n=6)
PluginFilter/noop/params:_small-32                        520.6n ± 1%   504.5n ± 1%  -3.10% (p=0.002 n=6)
PluginFilter/noop/params:_real-32                         517.6n ± 1%   504.6n ± 1%  -2.53% (p=0.002 n=6)
PluginFilter/test/params:_small-32                        7.113µ ± 1%   7.148µ ± 2%       ~ (p=0.310 n=6)
PluginFilter/test/params:_real-32                         116.4µ ± 1%   116.3µ ± 2%       ~ (p=0.937 n=6)
PluginScore/noop-wat/params:_small-32                     322.0n ± 1%   315.5n ± 1%  -2.00% (p=0.002 n=6)
PluginScore/noop-wat/params:_real-32                      323.7n ± 4%   319.2n ± 1%  -1.41% (p=0.017 n=6)
PluginScore/noop/params:_small-32                         589.4n ± 2%   577.6n ± 1%       ~ (p=0.093 n=6)
PluginScore/noop/params:_real-32                          531.0n ± 1%   518.1n ± 1%  -2.42% (p=0.002 n=6)
PluginScore/test/params:_small-32                         3.876µ ± 3%   3.778µ ± 0%  -2.52% (p=0.002 n=6)
PluginScore/test/params:_real-32                          49.20µ ± 2%   49.80µ ± 1%  +1.22% (p=0.015 n=6)
PluginPrefilterFilterAndScore/noop-wat/params:_small-32   491.0n ± 1%   482.5n ± 0%  -1.73% (p=0.002 n=6)
PluginPrefilterFilterAndScore/noop-wat/params:_real-32    491.5n ± 0%   483.9n ± 1%  -1.57% (p=0.002 n=6)
PluginPrefilterFilterAndScore/noop/params:_small-32       818.5n ± 1%   821.6n ± 1%       ~ (p=0.180 n=6)
PluginPrefilterFilterAndScore/noop/params:_real-32        813.4n ± 1%   817.0n ± 2%       ~ (p=0.197 n=6)
PluginPrefilterFilterAndScore/test/params:_small-32       8.396µ ± 0%   8.537µ ± 1%  +1.68% (p=0.004 n=6)
PluginPrefilterFilterAndScore/test/params:_real-32        118.6µ ± 2%   119.9µ ± 2%       ~ (p=0.240 n=6)
geomean                                                   1.749µ        1.734µ       -0.90%
```