@codefromthecrypt 2023-07-05
sorry, there's a tidy related concern. If you can run `make check` and then squash + force-push for merge, much obliged.