@pelletier 2023-07-05
Odd I thought I this would allow you to push changes:

<img width="970" alt="image" src="https://github.com/kubernetes-sigs/kube-scheduler-wasm-extension/assets/172804/108a6ffd-a49e-4a69-872b-3296bdc6b2da">

Pushed the make check commits (not sure why there is a diff on go.mod 🤷🏻 )!