@codefromthecrypt 2023-07-05
@pelletier yeah no idea.. at first I wondered if main was a special case. Anyway thanks for the help!