@codefromthecrypt 2023-07-05
thanks for the help, this will be handy again as we look at impact of passing state around 🥇